requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
# awscli
flake8
python-dotenv>=0.5.1

# custom requirements
ipykernel
jupyterlab
psycopg2-binary
sqlalchemy
pandas
matplotlib
seaborn
fastparquet
pyarrow
openpyxl
dvc
evidently
tzlocal
mlflow
xgboost
umap-learn
imbalanced-learn